requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy==1.24.3
faiss-cpu==1.7.4
requests==2.31.0
python-dotenv==1.0.0
spacy==3.7.2
scikit-learn==1.3.2
